+++ src/main/java/KUSITMS/WITHUS/domain/user/user/service/UserService.java
@@ -11,9 +11,4 @@ public interface UserService {
     void userJoinProcess(UserRequestDTO.UserJoin request);
     void resetPassword(String email, String newPassword);
     boolean isEmailDuplicated(String email);
-    void requestEmailVerification(String name, String email);
-    void requestPhoneVerification(String phoneNumber);
-    void confirmVerification(String identifier, String inputCode);
-    String reissueAccessToken(String refreshToken);
-    void logout();
 }
